src/test/java/interviews/google/_947_Most_Stones_Removed_with_Same_Row_or_Column/SolutionITest.java
package interviews.google._947_Most_Stones_Removed_with_Same_Row_or_Column;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 3/21/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final int[][] stones = {{0, 0}, {0, 1}, {1, 0}, {1, 2}, {2, 1}, {2, 2}};
    assertEquals(5, solution.removeStones(stones));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final int[][] stones = {{0, 0}, {0, 2}, {1, 1}, {2, 0}, {2, 2}};
    assertEquals(3, solution.removeStones(stones));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase3(Solution solution) {
    final int[][] stones = {{0, 0}};
    assertEquals(0, solution.removeStones(stones));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}